// src/app/api/generated/model/accept-invitacion-request.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Aceptacion de una invitacion, con los datos de alta si hacen falta
 */
export interface AcceptInvitacionRequest { 
    /**
     * Apellido. Opcional
     */
    apellido?: string;
    /**
     * Nombre. Obligatorio solo si el invitado no tiene cuenta todavia
     */
    nombre?: string;
    /**
     * Contrasena elegida. Obligatoria solo si el invitado no tiene cuenta todavia. Misma politica que el registro
     */
    password?: string;
    /**
     * Token recibido por correo
     */
    token: string;
}