// src/app/api/generated/model/invitacion-preview-response.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Lo que el invitado necesita saber para decidir
 */
export interface InvitacionPreviewResponse { 
    /**
     * Sede del vinculo, o null si es de toda la organizacion
     */
    consultorioNombre?: string;
    /**
     * Direccion a la que se emitio
     */
    email?: string;
    /**
     * Hasta cuando sirve el enlace
     */
    expiraEn?: string;
    /**
     * Organizacion que invita
     */
    organizacionNombre?: string;
    /**
     * Si al aceptar hay que crear la cuenta
     */
    requiereRegistro?: boolean;
    roleCode?: string;
}